+++ src/locales/en.json
@@ -226,7 +226,6 @@
       "code_hash": "Code Hash: ",
       "args": "Args: ",
       "hash_type": "Hash Type: ",
-      "assets": "Assets",
       "overview": "Overview",
       "user_defined_token": "User Defined Token",
       "confirmation": "Confirmation",
