+++ src/locales/zh.json
@@ -225,7 +225,6 @@
       "code_hash": "Code Hash: ",
       "args": "Args: ",
       "hash_type": "Hash Type: ",
-      "assets": "资产",
       "overview": "概览",
       "user_defined_token": "User Defined Token",
       "confirmation": "确认区块",
