+++ src/pages/Address/AddressComp.tsx
@@ -1,6 +1,5 @@
 import axios, { AxiosResponse } from 'axios'
-import { useState, useEffect, FC, ReactNode } from 'react'
-import { useHistory } from 'react-router'
+import { useState, useEffect, FC } from 'react'
 import { useQuery } from 'react-query'
 import { Radio } from 'antd'
 import Pagination from '../../components/Pagination'
@@ -11,6 +10,8 @@ import i18n from '../../utils/i18n'
 import { localeNumberString, parseUDTAmount } from '../../utils/number'
 import { shannonToCkb, deprecatedAddrToNewAddr, handleNftImgError, patchMibaoImg } from '../../utils/util'
 import {
+  AddressLockScriptController,
+  AddressLockScriptPanel,
   AddressTransactionsPagination,
   AddressTransactionsPanel,
   AddressUDTAssetsPanel,
@@ -21,9 +22,24 @@ import TitleCard from '../../components/Card/TitleCard'
 import CKBTokenIcon from '../../assets/ckb_token_icon.png'
 import SUDTTokenIcon from '../../assets/sudt_token.png'
 import { sliceNftName } from '../../utils/string'
-import { useIsLGScreen, useIsMobile, useNewAddr, useSearchParams } from '../../utils/hook'
+import {
+  useIsLGScreen,
+  useIsMobile,
+  useNewAddr,
+  usePaginationParamsInListPage,
+  useSearchParams,
+  useUpdateSearchParams,
+} from '../../utils/hook'
 import styles from './styles.module.scss'
 import TransactionLiteItem from '../../components/TransactionItem/TransactionLiteItem'
+import Script from '../../components/Script'
+import AddressText from '../../components/AddressText'
+import { parseSimpleDateNoSecond } from '../../utils/date'
+import { isMainnet } from '../../utils/chain'
+import ArrowUpIcon from '../../assets/arrow_up.png'
+import ArrowUpBlueIcon from '../../assets/arrow_up_blue.png'
+import ArrowDownIcon from '../../assets/arrow_down.png'
+import ArrowDownBlueIcon from '../../assets/arrow_down_blue.png'
 
 const addressAssetInfo = (address: State.Address, useMiniStyle: boolean) => {
   const items = [
@@ -166,7 +182,68 @@ interface CoTAList {
   }
 }
 
-export const AddressAssetComp: FC<{ address: State.Address; children?: ReactNode }> = ({ address, children }) => {
+const lockScriptIcon = (show: boolean) => {
+  if (show) {
+    return isMainnet() ? ArrowUpIcon : ArrowUpBlueIcon
+  }
+  return isMainnet() ? ArrowDownIcon : ArrowDownBlueIcon
+}
+
+const getAddressInfo = ({ liveCellsCount, minedBlocksCount, type, addressHash, lockInfo }: State.Address) => {
+  const items: OverviewItemData[] = [
+    {
+      title: i18n.t('address.live_cells'),
+      content: localeNumberString(liveCellsCount),
+    },
+    {
+      title: i18n.t('address.block_mined'),
+      content: localeNumberString(minedBlocksCount),
+    },
+  ]
+
+  if (type === 'LockHash') {
+    if (!addressHash) {
+      items.push({
+        title: i18n.t('address.address'),
+        content: i18n.t('address.unable_decode_address'),
+      })
+    } else {
+      items.push({
+        title: i18n.t('address.address'),
+        contentWrapperClass: styles.addressWidthModify,
+        content: <AddressText>{addressHash}</AddressText>,
+      })
+    }
+  }
+  if (lockInfo && lockInfo.epochNumber !== '0' && lockInfo.estimatedUnlockTime !== '0') {
+    const estimate = Number(lockInfo.estimatedUnlockTime) > new Date().getTime() ? i18n.t('address.estimated') : ''
+    items.push({
+      title: i18n.t('address.lock_until'),
+      content: `${lockInfo.epochNumber} ${i18n.t('address.epoch')} (${estimate} ${parseSimpleDateNoSecond(
+        lockInfo.estimatedUnlockTime,
+      )})`,
+    })
+  }
+  return items
+}
+
+const AddressLockScript: FC<{ address: State.Address }> = ({ address }) => {
+  const [showLock, setShowLock] = useState<boolean>(false)
+
+  return (
+    <AddressLockScriptPanel>
+      <OverviewCard items={getAddressInfo(address)} hideShadow>
+        <AddressLockScriptController onClick={() => setShowLock(!showLock)}>
+          <div>{i18n.t('address.lock_script')}</div>
+          <img alt="lock script" src={lockScriptIcon(showLock)} />
+        </AddressLockScriptController>
+        {showLock && address.lockScript && <Script script={address.lockScript} />}
+      </OverviewCard>
+    </AddressLockScriptPanel>
+  )
+}
+
+export const AddressOverview: FC<{ address: State.Address }> = ({ address }) => {
   const isLG = useIsLGScreen()
   const { udtAccounts = [] } = address
 
@@ -218,46 +295,36 @@ export const AddressAssetComp: FC<{ address: State.Address; children?: ReactNode
           </div>
         </AddressUDTAssetsPanel>
       ) : null}
-      {children}
+      <AddressLockScript address={address} />
     </OverviewCard>
   )
 }
 
 export const AddressTransactions = ({
-  currentPage,
-  pageSize,
   address,
   transactions,
   transactionsTotal: total,
   addressInfo: { addressHash },
 }: {
-  currentPage: number
-  pageSize: number
   address: string
   transactions: State.Transaction[]
   transactionsTotal: number
   addressInfo: State.Address
 }) => {
-  const history = useHistory()
   const isMobile = useIsMobile()
+  const { currentPage, pageSize, setPage } = usePaginationParamsInListPage()
   const searchParams = useSearchParams('layout')
-  const layout = searchParams.layout === 'professional' ? 'professional' : 'lite'
+  const defaultLayout = 'professional'
+  const updateSearchParams = useUpdateSearchParams<'layout'>()
+  const layout = searchParams.layout === 'lite' ? 'lite' : defaultLayout
   const totalPages = Math.ceil(total / pageSize)
 
-  const handleChange = (page: number, lo: 'professional' | 'lite') => {
-    if (page === 1) {
-      history.replace(`/address/${address}${lo === 'lite' ? '?layout=lite' : ''}`)
-    } else {
-      history.replace(`/address/${address}?page=${page}&size=${pageSize}${lo === 'lite' ? '&layout=lite' : ''}`)
-    }
-  }
-
-  const onChange = (page: number) => {
-    handleChange(page, layout)
-  }
-
   const onChangeLayout = (lo: 'professional' | 'lite') => {
-    handleChange(currentPage, lo)
+    updateSearchParams(params =>
+      lo === defaultLayout
+        ? Object.fromEntries(Object.entries(params).filter(entry => entry[0] !== 'layout'))
+        : { ...params, layout: lo },
+    )
   }
 
   const newAddr = useNewAddr(address)
@@ -297,38 +364,37 @@ export const AddressTransactions = ({
         }
       />
       <AddressTransactionsPanel>
-        {layout === 'lite' && !isMobile && (
-          <div className={styles.liteTransactionHeader}>
-            <div>{i18n.t('transaction.transaction_hash')}</div>
-            <div>{i18n.t('transaction.height')}</div>
-            <div>{i18n.t('transaction.time')}</div>
-            <div>{`${i18n.t('transaction.input')} & ${i18n.t('transaction.output')}`}</div>
-            <div>{i18n.t('transaction.capacity_change')}</div>
-          </div>
+        {layout === 'lite' ? (
+          <>
+            {!isMobile && (
+              <div className={styles.liteTransactionHeader}>
+                <div>{i18n.t('transaction.transaction_hash')}</div>
+                <div>{i18n.t('transaction.height')}</div>
+                <div>{i18n.t('transaction.time')}</div>
+                <div>{`${i18n.t('transaction.input')} & ${i18n.t('transaction.output')}`}</div>
+                <div>{i18n.t('transaction.capacity_change')}</div>
+              </div>
+            )}
+            {txList.map((transaction: State.Transaction) => (
+              <TransactionLiteItem address={addressHash} transaction={transaction} key={transaction.transactionHash} />
+            ))}
+          </>
+        ) : (
+          txList.map((transaction: State.Transaction, index: number) => (
+            <TransactionItem
+              address={addressHash}
+              transaction={transaction}
+              key={transaction.transactionHash}
+              circleCorner={{
+                bottom: index === transactions.length - 1 && totalPages === 1,
+              }}
+            />
+          ))
         )}
-        {txList.map((transaction: State.Transaction, index: number) => {
-          const { transactionHash } = transaction
-          if (transaction && layout === 'professional') {
-            return (
-              <TransactionItem
-                address={addressHash}
-                transaction={transaction}
-                key={transactionHash}
-                circleCorner={{
-                  bottom: index === transactions.length - 1 && totalPages === 1,
-                }}
-              />
-            )
-          }
-          if (transaction && layout === 'lite') {
-            return <TransactionLiteItem address={addressHash} transaction={transaction} key={transactionHash} />
-          }
-          return null
-        })}
       </AddressTransactionsPanel>
       {totalPages > 1 && (
         <AddressTransactionsPagination>
-          <Pagination currentPage={currentPage} totalPages={totalPages} onChange={onChange} />
+          <Pagination currentPage={currentPage} totalPages={totalPages} onChange={setPage} />
         </AddressTransactionsPagination>
       )}
     </>
@@ -336,6 +402,6 @@ export const AddressTransactions = ({
 }
 
 export default {
-  AddressAssetComp,
+  AddressOverview,
   AddressTransactions,
 }
